requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
Authlib==1.6.0
beautifulsoup4==4.13.3
blinker==1.9.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cmake==3.31.6
cmdstanpy==1.2.5
contourpy==1.3.2
cryptography==45.0.4
curl_cffi==0.11.1
cycler==0.12.1
Flask==3.1.0
fonttools==4.58.0
frozendict==2.4.6
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-genai==1.20.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.73.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.16.0
holidays==0.75
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
listing_scraper @ git+https://github.com/KobaKhit/listing-scraper.git@daa74a572156adba7eb042547f9bf3542e82b7bc
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mpld3==0.5.10
msgpack==1.1.1
multitasking==0.0.11
narwhals==1.41.0
newsapi-python==0.2.7
nltk==3.9.1
numpy==2.2.3
oauthlib==3.2.2
opencv-python==4.11.0.86
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
pandas-datareader==0.10.0
pandas_ta==0.3.14b0
peewee==3.18.1
pillow==11.2.1
platformdirs==4.3.8
playwright==1.52.0
plotly==6.1.1
pluggy==1.6.0
prophet==1.1.7
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
pygame==2.6.1
pyparsing==3.2.3
PySocks==1.7.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytubefix==8.12.2
pytz==2025.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.15.2
selectolax==0.3.29
selenium==4.32.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stanio==0.5.1
threadpoolctl==3.5.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==1.26.20
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
yfinance==0.2.61
